% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appResultsLocal.R
\name{appResultsLocal}
\alias{appResultsLocal}
\title{appResultsLocal}
\usage{
appResultsLocal(folder = "output/", valfile = NULL)
}
\arguments{
\item{folder}{output folder containing the runs to be analyzed as subfolders (e.g. folder "output" in a
MAgPIE model folder)}

\item{valfile}{Path to a validation file, preferably in rds format, but can also be provided as mif (in the 
latter case it will be converted to rds first). If not path is given the function will look automatically for 
an validation file in the output folder}
}
\description{
version of appResults which is optimized to run on a local model folder. In contrast to appResults, 
appResultsLocal only requires the path to an output folder (with subfolders for each run).
}
\seealso{
\code{\link{appResults}}
}
\author{
Jan Philipp Dietrich, Lavinia Baumstark
}
