% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmarkState.R
\name{compressIds}
\alias{compressIds}
\title{Compress IDs for URL}
\usage{
compressIds(ids)
}
\arguments{
\item{ids}{Vector of numeric IDs}
}
\value{
Compressed string representation of IDs
}
\description{
Compresses a vector of numeric IDs into a shorter string representation
for use in URL query parameters.
}
\author{
Florian Humpenoeder
}
\keyword{internal}
