% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmarkState.R
\name{createBookmarkUrl}
\alias{createBookmarkUrl}
\title{Create Bookmark URL}
\usage{
createBookmarkUrl(session, state)
}
\arguments{
\item{session}{Shiny session object}

\item{state}{List containing application state to bookmark}
}
\value{
Full URL string with state encoded in query parameters
}
\description{
Creates a shareable URL with the current application state encoded.
}
\author{
Florian Humpenoeder
}
\keyword{internal}
