% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmarkState.R
\name{decodeAppState}
\alias{decodeAppState}
\title{Decode Application State}
\usage{
decodeAppState(str)
}
\arguments{
\item{str}{JSON string from URL}
}
\value{
List containing application state
}
\description{
Decodes a JSON string back into application state.
}
\author{
Florian Humpenoeder
}
\keyword{internal}
