% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmarkState.R
\name{decompressIds}
\alias{decompressIds}
\title{Decompress IDs from URL}
\usage{
decompressIds(str)
}
\arguments{
\item{str}{Compressed string from URL}
}
\value{
Vector of numeric IDs
}
\description{
Decompresses a string back into a vector of IDs.
}
\author{
Florian Humpenoeder
}
