% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmarkState.R
\name{encodeAppState}
\alias{encodeAppState}
\title{Encode Application State}
\usage{
encodeAppState(state)
}
\arguments{
\item{state}{List containing application state}
}
\value{
JSON string representation of state
}
\description{
Encodes the essential application state into a compact JSON string.
}
\author{
Florian Humpenoeder
}
\keyword{internal}
