% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadVariableConfig.R
\name{getPresetVariables}
\alias{getPresetVariables}
\title{Get Variables for Preset}
\usage{
getPresetVariables(config, presetId)
}
\arguments{
\item{config}{Variable configuration list from loadVariableConfig()}

\item{presetId}{ID of the preset to extract variables from}
}
\value{
List of variable configurations, each with 'variable' and 'plot_type' fields
}
\description{
Extract the list of variables from a specific preset.
}
\author{
Florian Humpenoeder
}
