% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modAreaPlot.R
\name{modAreaPlot}
\alias{modAreaPlot}
\title{modAreaPlot Module}
\usage{
modAreaPlot(id, report, selectionSets = NULL)
}
\arguments{
\item{id}{Module ID string, must match the id used in \code{\link{modAreaPlotUI}}}

\item{report}{A reactive containing the report to be visualized}

\item{selectionSets}{named list of selection sets per filter column (passed to \code{\link{modFilter}}).
Defaults to the \code{selectionSets} entry of the active \code{appResults} option.}
}
\description{
Shiny module which works together with \code{\link{modAreaPlotUI}} to produce an area plot tab
}
\seealso{
\code{\link{modAreaPlotUI}}, \code{\link{appResults}}
}
\author{
Jan Philipp Dietrich, Florian Humpenoeder
}
\keyword{internal}
