% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modAreaPlotUI.R
\name{modAreaPlotUI}
\alias{modAreaPlotUI}
\title{modAreaPlot Module}
\usage{
modAreaPlotUI(id)
}
\arguments{
\item{id}{id of the filter}
}
\description{
Shiny module which works together with \code{\link{modAreaPlot}} to produce an area plot tab
}
\seealso{
\code{\link{modAreaPlot}}, \code{\link{appResults}}
}
\author{
Florian Humpenoeder, Jan Philipp Dietrich
}
