% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modDashboard.R
\name{modDashboard}
\alias{modDashboard}
\title{modDashboard Server Module}
\usage{
modDashboard(input, output, session, report, validation, config)
}
\arguments{
\item{input, output, session}{Default input, output and session objects coming from shiny}

\item{report}{A reactive list containing the report data (from modRunSelect)}

\item{validation}{A reactive containing validation data to be shown}

\item{config}{Variable configuration list from loadVariableConfig()}
}
\description{
Shiny module which works together with \code{\link{modDashboardUI}} to produce a dashboard
with pre-configured plots based on variable presets.
}
\seealso{
\code{\link{modDashboardUI}}, \code{\link{appResults}}, \code{\link{loadVariableConfig}}
}
\author{
Florian Humpenoeder
}
