% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modDashboardUI.R
\name{modDashboardUI}
\alias{modDashboardUI}
\title{modDashboard UI Module}
\usage{
modDashboardUI(id, presets = NULL)
}
\arguments{
\item{id}{id of the module}

\item{presets}{Named list of presets from variable configuration}
}
\description{
Shiny module which works together with \code{\link{modDashboard}} to produce a dashboard
with pre-configured plots based on variable presets.
}
\seealso{
\code{\link{modDashboard}}, \code{\link{appResults}}, \code{\link{loadVariableConfig}}
}
\author{
Florian Humpenoeder
}
