% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modFilter.R
\name{modFilter}
\alias{modFilter}
\title{modFilter Module}
\usage{
modFilter(
  id,
  data,
  exclude = NULL,
  showAll = FALSE,
  multiple = NULL,
  xdata = NULL,
  xdataExclude = NULL,
  order = NULL,
  name = NULL,
  preselectYear = NULL,
  preselectMinDate = NULL,
  selectionSets = NULL
)
}
\arguments{
\item{id}{Module ID string, must match the id used in \code{\link{modFilterUI}}}

\item{data}{A reactive returning a data.table with observations in rows and filter options in columns}

\item{exclude}{names of columns that should be not used as filter}

\item{showAll}{FALSE | If set to TRUE all available filter are shown and the filter selector is hidden}

\item{multiple}{vector with booleans for each filter defining whether multiple selections are allowed
or not. If information is not provided it is assumed that multiple selection is allowed}

\item{xdata}{additional data.tables which should be filtered by the same rules as data. If provided
the format of the return value changes}

\item{xdataExclude}{similar to exclude a vector of filters that should be ignored for xdata. Useful if
xdata should only filtered for a subset of filters applied to data}

\item{order}{order the filter should be listed (provided as a vector of filter names). Filter not
listed here will be shown after the ones mentioned.}

\item{name}{name used to identify the filter in the log}

\item{preselectYear}{if provided the year filter will be preselected with this value}

\item{preselectMinDate}{if provided the date filter will be preselected with this as lower value}

\item{selectionSets}{named list of selection sets per filter column. Each element is a named list
mapping set labels to character vectors of values. When a set is selected, all its member values
are added to the selection. Only sets where all members are present in the data are shown.
Example: \code{list(region = list("All EUR regions" = c("EUC", "EUS", "EUW", "DEU")))}}
}
\value{
a reactive list with x as the filtered data and xdata containing the list of additional,
filtered data element.
}
\description{
Shiny module which works together with \code{\link{modFilterUI}} to filter a data set based on user input
}
\seealso{
\code{\link{modFilterUI}}, \code{\link{appModelstats}}
}
\author{
Jan Philipp Dietrich
}
\keyword{internal}
