% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modFilterUI.R
\name{modFilterUI}
\alias{modFilterUI}
\title{modFilterUI Module}
\usage{
modFilterUI(id)
}
\arguments{
\item{id}{id of the filter}
}
\description{
Corresponding user interface to \code{\link{modFilter}} to filter a data set based on user input
}
\seealso{
\code{\link{modFilter}}, \code{\link{appModelstats}}
}
\author{
Jan Philipp Dietrich
}
