% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modRunSelectUI.R
\name{modRunSelectUI}
\alias{modRunSelectUI}
\title{modRunSelectUI Module}
\usage{
modRunSelectUI(id)
}
\arguments{
\item{id}{id of the filter}
}
\description{
Corresponding user interface to \code{\link{modRunSelect}} to select modules runs for further analysis
}
\seealso{
\code{\link{modFilter}}, \code{\link{appModelstats}}
}
\author{
Jan Philipp Dietrich
}
