% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmarkState.R
\name{parseBookmarkUrl}
\alias{parseBookmarkUrl}
\title{Parse URL for Bookmarked State}
\usage{
parseBookmarkUrl(session)
}
\arguments{
\item{session}{Shiny session object}
}
\value{
List containing decoded application state, or empty list if none
}
\description{
Extracts and decodes application state from URL query parameters.
}
\author{
Florian Humpenoeder
}
