% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyresults-package.R
\docType{package}
\name{shinyresults-package}
\alias{shinyresults}
\alias{shinyresults-package}
\title{shinyresults: A collection of shiny apps and modules to visualize/handle model results}
\description{
A collection of shiny apps and modules to explore and visualize time series of modelling results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/shinyresults}
  \item \doi{10.5281/zenodo.1478922}
  \item Report bugs at \url{https://github.com/pik-piam/shinyresults/issues}
}

}
\author{
\strong{Maintainer}: Jan Philipp Dietrich \email{dietrich@pik-potsdam.de} (\href{https://orcid.org/0000-0002-4309-6431}{ORCID}) (Potsdam Institute for Climate Impact Research)

Authors:
\itemize{
  \item Florian Humpenoeder \email{humpenoeder@pik-potsdam.de} (Potsdam Institute for Climate Impact Research)
  \item Pascal Sauer \email{pascal.sauer@pik-potsdam.de} (\href{https://orcid.org/0000-0002-6856-8239}{ORCID}) (Potsdam Institute for Climate Impact Research)
  \item Patrick Rein \email{patrick.rein@pik-potsdam.de} (\href{https://orcid.org/0000-0001-9454-8381}{ORCID}) (Potsdam Institute for Climate Impact Research)
}

}
\keyword{internal}
