% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadVariableConfig.R
\name{validateVariableConfig}
\alias{validateVariableConfig}
\title{Validate Variable Configuration}
\usage{
validateVariableConfig(config)
}
\arguments{
\item{config}{List read from YAML file}
}
\value{
Validated and normalized config list
}
\description{
Validates the structure of a variable configuration list.
}
\author{
Florian Humpenoeder
}
\keyword{internal}
