% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modAreaPlot.R
\name{modAreaPlot}
\alias{modAreaPlot}
\title{modAreaPlot Module}
\usage{
modAreaPlot(input, output, session, report)
}
\arguments{
\item{input, output, session}{Default input, output and session objects coming from shiny}

\item{report}{A reactive containing the report to be visualized}
}
\description{
Shiny module which works together with \code{\link{modAreaPlotUI}} to produce an area plot tab
}
\seealso{
\code{\link{modAreaPlotUI}}, \code{\link{appResults}}
}
\author{
Jan Philipp Dietrich, Florian Humpenoeder
}
