% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modLinePlot.R
\name{modLinePlot}
\alias{modLinePlot}
\title{modLinePlot Module}
\usage{
modLinePlot(input, output, session, report, validation)
}
\arguments{
\item{input, output, session}{Default input, output and session objects coming from shiny}

\item{report}{A reactive containing the report to be visualized}

\item{validation}{A reactive containing validation data to be shown}
}
\description{
Shiny module which works together with \code{\link{modLinePlotUI}} to produce a line plot tab
}
\seealso{
\code{\link{modLinePlotUI}}, \code{\link{appResults}}
}
\author{
Florian Humpenoeder, Jan Philipp Dietrich
}
