% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modLinePlotUI.R
\name{modLinePlotUI}
\alias{modLinePlotUI}
\title{modLinePlot Module}
\usage{
modLinePlotUI(id)
}
\arguments{
\item{id}{id of the filter}
}
\description{
Shiny module which works together with \code{\link{modLinePlot}} to produce a line plot tab
}
\seealso{
\code{\link{modLinePlot}}, \code{\link{appResults}}
}
\author{
Florian Humpenoeder, Jan Philipp Dietrich
}
