% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyresults-package.R
\docType{package}
\name{shinyresults-package}
\alias{shinyresults}
\alias{shinyresults-package}
\title{shinyresults: A collection of shiny apps and modules to visualize/handle model results}
\description{
A collection of tools which allow to manipulate and analyze code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/shinyresults}
  \item \doi{10.5281/zenodo.1478922}
  \item Report bugs at \url{https://github.com/pik-piam/shinyresults/issues}
}

}
\author{
\strong{Maintainer}: Jan Philipp Dietrich \email{dietrich@pik-potsdam.de}

Authors:
\itemize{
  \item Florian Humpenoeder \email{humpenoeder@pik-potsdam.de}
}

}
